requirements.txt
openai>=1.13
chromadb>=0.4
PyMuPDF>=1.23
streamlit>=1.30
tiktoken>=0.5
plotly>=5.0
pandas>=2.0
python-dotenv>=1.0
tqdm>=4.65
langchain-text-splitters>=0.0.1
pypdf>=4.0
